backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.1
PyGithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
